refactor: remove barel files
