+++ tdp/cli/commands/deploy.py
@@ -9,20 +9,19 @@ from typing import TYPE_CHECKING
 import click
 from sqlalchemy import Engine
 
-from tdp.cli.params import (
-    collections_option,
-    database_dsn_option,
-    validate_option,
-    vars_option,
-)
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.database_dsn_option import database_dsn_option
+from tdp.cli.params.validate_option import validate_option
+from tdp.cli.params.vars_option import vars_option
 from tdp.cli.utils import check_services_cleanliness
-from tdp.core.deployment import DeploymentRunner, Executor
+from tdp.core.deployment.deployment_runner import DeploymentRunner
+from tdp.core.deployment.executor import Executor
 from tdp.core.models.enums import DeploymentStateEnum
-from tdp.core.variables import ClusterVariables
+from tdp.core.variables.cluster_variables import ClusterVariables
 from tdp.dao import Dao
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 
 @click.command()
