+++ tdp/cli/commands/plan/import_file.py
@@ -9,14 +9,15 @@ from typing import TYPE_CHECKING
 import click
 from sqlalchemy import Engine
 
-from tdp.cli.params import collections_option, database_dsn_option
-from tdp.cli.params.plan import force_option
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.database_dsn_option import database_dsn_option
+from tdp.cli.params.plan.force_option import force_option
 from tdp.cli.utils import parse_file
 from tdp.core.models.deployment_model import DeploymentModel
 from tdp.dao import Dao
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 logger = logging.getLogger(__name__)
