+++ tdp/cli/commands/plan/reconfigure.py
@@ -8,16 +8,17 @@ from typing import TYPE_CHECKING, Optional
 import click
 from sqlalchemy import Engine
 
-from tdp.cli.params import collections_option, database_dsn_option
-from tdp.cli.params.plan import force_option, preview_option, rolling_interval_option
-from tdp.cli.utils import (
-    print_deployment,
-)
-from tdp.core.models import DeploymentModel
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.database_dsn_option import database_dsn_option
+from tdp.cli.params.plan.force_option import force_option
+from tdp.cli.params.plan.preview_option import preview_option
+from tdp.cli.params.plan.rolling_interval_option import rolling_interval_option
+from tdp.cli.utils import print_deployment
+from tdp.core.models.deployment_model import DeploymentModel
 from tdp.dao import Dao
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 
 @click.command()
