+++ tdp/cli/commands/plan/resume.py
@@ -8,14 +8,15 @@ from typing import TYPE_CHECKING, Optional
 import click
 from sqlalchemy import Engine
 
-from tdp.cli.params import collections_option, database_dsn_option
-from tdp.cli.params.plan import preview_option
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.database_dsn_option import database_dsn_option
+from tdp.cli.params.plan.preview_option import preview_option
 from tdp.cli.utils import print_deployment
-from tdp.core.models import DeploymentModel
+from tdp.core.models.deployment_model import DeploymentModel
 from tdp.dao import Dao
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 
 @click.command()
