+++ tdp/cli/commands/status/show.py
@@ -9,29 +9,25 @@ import click
 from sqlalchemy import Engine
 from tabulate import tabulate
 
-from tdp.cli.params import (
-    collections_option,
-    database_dsn_option,
-    hosts_option,
-    validate_option,
-    vars_option,
-)
-from tdp.cli.params.status import (
-    component_argument_option,
-    service_argument_option,
-)
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.database_dsn_option import database_dsn_option
+from tdp.cli.params.hosts_option import hosts_option
+from tdp.cli.params.status.component_argument import component_argument_option
+from tdp.cli.params.status.service_argument import service_argument_option
+from tdp.cli.params.validate_option import validate_option
+from tdp.cli.params.vars_option import vars_option
 from tdp.cli.utils import (
     check_services_cleanliness,
     print_hosted_entity_status_log,
 )
 from tdp.core.models.sch_status_log_model import SCHStatusLogModel
-from tdp.core.variables import ClusterVariables
+from tdp.core.variables.cluster_variables import ClusterVariables
 from tdp.dao import Dao
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 
 def _filter_stale(stale: Optional[bool], no_stale: Optional[bool]) -> Optional[bool]:
