+++ tdp/cli/commands/vars/validate.py
@@ -8,11 +8,12 @@ from typing import TYPE_CHECKING
 
 import click
 
-from tdp.cli.params import collections_option, vars_option
-from tdp.core.variables import ClusterVariables
+from tdp.cli.params.collections_option import collections_option
+from tdp.cli.params.vars_option import vars_option
+from tdp.core.variables.cluster_variables import ClusterVariables
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
 
 
 @click.command()
