+++ tdp/cli/params/status/__init__.py
@@ -1,5 +1,2 @@
 # Copyright 2022 TOSIT.IO
 # SPDX-License-Identifier: Apache-2.0
-
-from .component_argument import component_argument_option
-from .service_argument import service_argument_option
