+++ tdp/core/collections/__init__.py
@@ -1,4 +1,2 @@
 # Copyright 2022 TOSIT.IO
 # SPDX-License-Identifier: Apache-2.0
-
-from .collections import Collections
