+++ tdp/core/deployment/__init__.py
@@ -1,6 +1,2 @@
 # Copyright 2022 TOSIT.IO
 # SPDX-License-Identifier: Apache-2.0
-
-from tdp.core.deployment.deployment_iterator import DeploymentIterator
-from tdp.core.deployment.deployment_runner import DeploymentRunner
-from tdp.core.deployment.executor import Executor
