+++ tdp/core/models/deployment_model.py
@@ -26,7 +26,7 @@ from tdp.core.models.enums import (
 from tdp.core.models.operation_model import OperationModel
 
 if TYPE_CHECKING:
-    from tdp.core.collections import Collections
+    from tdp.core.collections.collections import Collections
     from tdp.core.entities.hosted_entity_status import HostedEntityStatus
     from tdp.core.entities.operation import Operation
