+++ tdp/core/models/operation_model.py
@@ -18,7 +18,7 @@ from tdp.core.models.base_model import BaseModel
 from tdp.core.models.enums import OperationStateEnum
 
 if TYPE_CHECKING:
-    from tdp.core.models import DeploymentModel
+    from tdp.core.models.deployment_model import DeploymentModel
 
 
 class OperationModel(BaseModel):
