+++ tdp/core/variables/schema/__init__.py
@@ -1,12 +1,2 @@
 # Copyright 2022 TOSIT.IO
 # SPDX-License-Identifier: Apache-2.0
-
-from tdp.core.variables.schema.exceptions import (
-    InvalidSchemaError,
-    InvalidVariablesError,
-    SchemaValidationError,
-)
-from tdp.core.variables.schema.service_collection_schema import ServiceCollectionSchema
-from tdp.core.variables.schema.service_schema import ServiceSchema
-from tdp.core.variables.schema.types import Schema
-from tdp.core.variables.schema.validator import VariablesValidator
