+++ tests/conftest.py
@@ -18,7 +18,7 @@ from tdp.core.constants import (
     PLAYBOOKS_DIRECTORY_NAME,
     YML_EXTENSION,
 )
-from tdp.core.models import BaseModel, init_database
+from tdp.core.models.base_model import BaseModel, init_database
 
 
 def pytest_addoption(parser: pytest.Parser) -> None:
@@ -35,7 +35,9 @@ def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
     """Pytest hook to generate tests based on the database dsn option."""
     if "db_dsn" in metafunc.fixturenames:
         metafunc.parametrize(
-            "db_dsn", metafunc.config.getoption("database_dsn"), indirect=True  # type: ignore
+            "db_dsn",
+            metafunc.config.getoption("database_dsn"),
+            indirect=True,  # type: ignore
         )
